templating module
Overrides jinjas default variable_start_string
and variable_end_string to avoid collisions with
valid ruby code.

Uses '|{' for start and '}|' for close.

Includes a generator function for bulk rendering
templates from original paths on the fly when they
all need to be rendered from the same map. The generator
was used because the result hold the entire contents
of the rendered file.

A non-generator is also available.