update MetadataRb implementation
The CookBook objects 'metadata' property, the
MetadataRb instance now parallels the Berksfile
instance in the 'berksfile' property. This includes
a merge() method and a depends_statment() method
which is used by parse() and from_dict() for translating
the files to and from python dictionaries.

Note that the conversion is not a perfectly symmetrical
translation (yet).

food.update_cookbook() now returns a tuple of
written_files, cookbook. This will also be used
in functional tests to veryify the outcome of gen/build.

CookBook now has a name property which is read from the
cookbook's metadata.rb